requirements.txt
convokit~=3.0.0
flake8~=7.1.1
jsonlines~=4.0.0
matplotlib~=3.9.2
nltk~=3.9.1
numpy~=1.26.4
openai~=1.47.0
pandas~=2.2.3
python-dotenv~=1.0.1
scikit-learn~=1.5.2
torch~=2.1.1
tqdm~=4.66.5
